Добавлена документация, исправлена ошибки парсинга

+++ diploma_v2/Form2.cs
@@ -23,8 +23,8 @@ namespace diploma_v2
             {
                 Program.task.dimX = Convert.ToInt32(textBox1.Text, Program.nfi);
                 Program.task.dimP = Convert.ToInt32(textBox2.Text, Program.nfi);
-                Program.task.t0 = Convert.ToInt32(textBox3.Text, Program.nfi);
-                Program.task.T = Convert.ToInt32(textBox4.Text, Program.nfi);
+                Program.task.t0 = Convert.ToDouble(textBox3.Text, Program.nfi);
+                Program.task.T = Convert.ToDouble(textBox4.Text, Program.nfi);
                 Program.task.dimTk = Convert.ToInt32(textBox5.Text, Program.nfi);
                 Program.task.step = Convert.ToDouble(textBox6.Text, Program.nfi);
                 Program.task.m = radioButton2.Checked;
